[issue-9] make ColorScheme the representation
